test_runner: read `then` exactly once when coercing a thenable
ResolvePromise already fulfills a non-thenable synchronously, so the
promise's status after resolve() says whether the value was a thenable.
Use that instead of a separate `then` lookup, which made a `then` getter
run twice.

Move the thenable cases into the existing test-test.test.ts suite.
